[NextDNS] solve more merge conflicts
