+++ extensions/nextdns/src/components/list-item.tsx
@@ -1,18 +1,44 @@
 import { Action, ActionPanel, Color, Icon, List, showToast, Toast } from "@raycast/api";
-import { toggleSite } from "../libs/api";
-import { Mutate } from "../types/nextdns";
-import AddSite from "./add-site";
+import { toggleDomain } from "../libs/api";
+import { DomainListItem, Mutate } from "../types/nextdns";
+import AddDomain from "./add-domain";
 
 //TODO: Optimize optimistic update
 //TODO: Ensure naming, site or domain?
-export function ListItem(props: { id: string; active: boolean; type: string, mutate: Mutate}) {
-  const { id, active, type, mutate } = props;
+export function ListItem(props: { siteItem: DomainListItem, type: "allow" | "deny", mutate: Mutate }) {
+  const { siteItem, type, mutate } = props;
 
-  async function toggle(newStatus: boolean) {
+  return (
+    <List.Item
+      title={`*.${siteItem.id}`}
+      icon={
+        siteItem.active
+          ? { source: Icon.CheckCircle, tintColor: Color.Green }
+          : { source: Icon.Circle, tintColor: Color.SecondaryText }
+      }
+      actions={<Actions item={siteItem} type={type} mutate={mutate} />}
+    />
+  );
+}
+
+
+function Actions({
+  item,
+  type,
+  mutate
+}: {
+  item: DomainListItem;
+  type: "allow" | "deny";
+  mutate: Mutate
+}) {
+  async function toggle(item: DomainListItem) {
+    const newStatus = !item.active;
+    const { id, type } = item;
+
     const toast = await showToast({ style: Toast.Style.Animated, title: newStatus ? "Activating Site" : "Deactivating Site" });
     try {
       await mutate(
-        toggleSite({ id, type, active: newStatus }),
+        toggleDomain({ element: {id, type, active: newStatus} }),
         {
           optimisticUpdate(data) {
             const { result, profileName } = data;
@@ -31,36 +57,13 @@ export function ListItem(props: { id: string; active: boolean; type: string, mut
   }
 
   return (
-    <List.Item
-      title={`*.${id}`}
-      icon={
-        siteItem.active
-          ? { source: Icon.CheckCircle, tintColor: Color.Green }
-          : { source: Icon.Circle, tintColor: Color.SecondaryText }
-      }
-      actions={<Actions item={siteItem} onItemRemove={onRemoveItem} />}
-    />
-  );
-}
-
-
-function Actions({
-  item,
-  onItemRemove,
-}: {
-  item: DomainListItem;
-  onItemRemove: (item: DomainListItem) => Promise<void>;
-}) {
-  return (
     <ActionPanel title={`Manage ${item.id}`}>
       <Action
         title={`${item.active ? "Deactivate" : "Activate"} Domain`}
         icon={item.active ? Icon.XMarkCircle : Icon.CheckCircle}
-        onAction={async () => {
-          await onItemRemove(item);
-        }}
+        onAction={() => toggle(item)}
       />
-      <Action.Push target={<AddDomain />} title="Push" />
+      <Action.Push icon={Icon.Plus} title="Add New Site" target={<AddDomain type={type} mutate={mutate} />} />
     </ActionPanel>
   );
 }
