+++ extensions/nextdns/src/denylist.tsx
@@ -1,7 +1,5 @@
 import { getDomains } from "./libs/api";
 import { ErrorView } from "./components/error-view";
-import { DomainListItem } from "./types/nextdns";
-import { removeItem } from "./libs/utils";
 import { DomainList } from "./libs/domainList";
 
 export default function Command() {
@@ -12,21 +10,7 @@ export default function Command() {
     return <ErrorView />;
   }
 
-return (
-    <List
-      isLoading={isLoading}
-      searchBarPlaceholder={`Search denylist of ${data?.profileName ?? "..."} (${PREFERENCES.nextdns_profile_id})`}
-    >
-      {data && (
-        <>
-          {data.result?.map((element) => <ListItem key={element.id} id={element.id} active={element.active} type="deny" mutate={mutate} />)}
-
-          {Object.keys(data).length === 0 && (
-            <EmptyView title="No domains in denylist" icon={{ source: "no_view.png" }} />
-          )}
-        </>
-      )}
-      <EmptyView title="No Results" icon={{ source: "no_view.png" }} />
-    </List>
+  return (
+    <DomainList data={data} type={type} isLoading={isLoading} mutate={mutate} />
   );
 }
