+++ extensions/nextdns/src/libs/domainList.tsx
@@ -4,16 +4,16 @@ import { EmptyView } from "../components/empty-view";
 import { ListItem } from "../components/list-item";
 import { PREFERENCES } from "./constants";
 
-export const DomainList: React.FC<DomainListProps> = ({ data, isLoading, onRemoveItem }) => (
+export const DomainList: React.FC<DomainListProps> = ({ data, type, isLoading, mutate}) => (
   <List
     isLoading={isLoading}
-    searchBarPlaceholder={`Search allowlist of ${data.profileName} (${PREFERENCES.nextdns_profile_id}`}
+    searchBarPlaceholder={`Search ${type}list of ${data.profileName} (${PREFERENCES.nextdns_profile_id}`}
   >
     {data.result?.map((element: DomainListItem) => (
-      <ListItem key={element.id} siteItem={element} onRemoveItem={onRemoveItem} />
+      <ListItem key={element.id} siteItem={element} type={type} mutate={mutate} />
     ))}
 
-    {Object.keys(data).length === 0 && <EmptyView title="No domains in allowlist" icon={{ source: "no_view.png" }} />}
+    {Object.keys(data).length === 0 && <EmptyView title={`No domains in ${type}list`} icon={{ source: "no_view.png" }} />}
     <EmptyView title="No Results" icon={{ source: "no_view.png" }} />
   </List>
 );
